feat: 서비스 로직 분리(#34)
